Rename : 컴포넌트 위치 이동
상위 컴포넌트 하위 컴포넌트 구분 적용
