+++ finda2.0/src/components/Main/MainTop/SearchBar/Index.style.ts
