+++ finda2.0/src/components/Movie/Offer/Index.style.ts
