+++ finda2.0/src/components/Movie/Offer/type.ts
