+++ finda2.0/src/components/Result/NoResult/Index.style.ts
