+++ finda2.0/src/components/Result/NoResult/type.ts
